requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.6b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.1
cloudinary==1.28.0
colorama==0.4.4
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.8.1
django-taggit==2.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.3.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.1
regex==2021.7.6
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
waitress==2.0.0
whitenoise==5.3.0